requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
kafka-python==2.3.0
PyYAML==6.0.1
jsonpath-ng==1.6.1
pydantic==2.5.0
pydantic-settings==2.1.0
six==1.17.0
fastavro==1.8.4
requests==2.31.0
